separates the objects from where the menus and submenus should be named to another file
